add storage variable caching initial impl
