+++ examples/optimized_contracts/storage_variable_caching.sol
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: MIT
+pragma solidity 0.8.20;
+
+contract Counter1 {
+    uint256 public number;
+
+    function increment() public {
+        uint256 _number = number;
+        require(_number < 10);
+        number = _number + 1;
+    }
+}
