+++ examples/unoptimized_contracts/storage_variable_caching.sol
@@ -1,0 +1,11 @@
+// SPDX-License-Identifier: MIT
+pragma solidity 0.8.20;
+
+contract Counter1 {
+    uint256 public number;
+
+    function increment() public {
+        require(number < 10);
+        number = number + 1;
+    }
+}
